+++ selectrum.el
@@ -434,7 +434,7 @@ destructively and return the modified list."
         (setq link (cdr link))))
     (nconc (nreverse elts) (cdr lst))))
 
-(defun selectrum--normalize-collection (collection &optional predicate)
+(defun selectrum--normalize-collection (collection &optional predicate buffer)
   "Normalize COLLECTION into a list of strings.
 COLLECTION may be a list of strings or symbols or cons cells, an
 obarray, a hash table, or a function, as per the docstring of
@@ -442,9 +442,18 @@ obarray, a hash table, or a function, as per the docstring of
 damaging the original COLLECTION.
 
 If PREDICATE is non-nil, then it filters the collection as in
-`all-completions'."
-  (let ((completion-regexp-list nil))
-    (all-completions "" collection predicate)))
+`all-completions'.
+
+BUFFER is the buffer to compute the candidates in. It defaults to
+buffer of the selected window except for minibuffers where it
+defaults to the buffer of `minibuffer-selected-window'."
+  ;; Making the last buffer current avoids the cost of potential
+  ;; buffer switching for each candidate within the predicate (see
+  ;; `describe-variable').
+  (with-current-buffer (or buffer
+                           (window-buffer (minibuffer-selected-window)))
+    (let ((completion-regexp-list nil))
+      (all-completions "" collection predicate))))
 
 (defun selectrum--remove-default-from-prompt (prompt)
   "Remove the indication of the default value from PROMPT.
@@ -681,13 +690,12 @@ Window will be created by `selectrum-display-action'."
                    (setq buffer-read-only t)
                    (setq show-trailing-whitespace nil)
                    (goto-char (point-min))
-                   (current-buffer)))))
+                   (current-buffer))))
+        (action selectrum-display-action))
     (or (get-buffer-window buf 'visible)
         (with-selected-window (minibuffer-selected-window)
           (let* ((frame (selected-frame))
-                 (window (display-buffer
-                          buf
-                          selectrum-display-action)))
+                 (window (display-buffer buf action)))
             (select-frame-set-input-focus frame)
             window)))))
 
@@ -994,18 +1002,17 @@ The specific details of the formatting are determined by
                                (minibuffer-contents)
                                lines)))
                  (match
-                  (concat
-                   (propertize
-                    (propertize newline/display 'face newline/face)
-                    'selectrum-candidate-display-prefix
-                    (number-to-string (1- len)))
+                  (propertize
                    (replace-regexp-in-string
                     "[ \t][ \t]+"
                     (propertize whitespace/display 'face whitespace/face)
                     (if (string-empty-p (minibuffer-contents))
                         ""
                       ;; Show first matched line.
-                      (or fmatch "")) 'fixed-case 'literal)))
+                      (or fmatch "")) 'fixed-case 'literal)
+                   'selectrum-candidate-display-prefix
+                   (propertize (format "(%d lines)" len)
+                               'face newline/face)))
                  (annot (replace-regexp-in-string
                          "\n" (propertize newline/display 'face newline/face)
                          (replace-regexp-in-string